# backend/swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0",
    "description": "Work with contracting companies in the field of maintenance and repair of buildings, communications and equipment.",
    "title": "Order service"
  },
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/refresh": {
      "post": {
        "summary": "refresh token",
        "description": "",
        "operationId": "refresh",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "Success response",
            "schema": {
              "type": "object",
              "properties": {
                "accessToken": {
                  "type": "string",
                  "example": "GciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImRlbnlrLnlldmhlbkBnbWFpbC5jb20iLCJpYXQiOjE2ODE3NTM5MTUs"
                },
                "refreshToken": {
                  "type": "string",
                  "example": "GciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImRlbnlrLnlldmhlbkBnbWFpbC5jb20iLCJpYXQiOjE2ODE3NTM5MTUs"
                }
              }
            }
          },
          "401": {
            "description": "No token / Unknown token / Token invalid"
          }
        }
      }
    },
...
    "/auth//password/forgot": {
      "post": {
        "summary": "forgot password",
        "description": "receive an email to reset your password",
        "operationId": "ForgotPassword",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "email",
            "type": "string",
            "required": true
          },
          {
            "in": "formData",
            "name": "contractor",
            "type": "boolean",
            "description": "if you are a contractor"
          }
        ],
        "responses": {
          "200": {
            "description": "The letter was send to the your email"
          },
          "400": {
            "description": "Body invalid"
          },
          "404": {
            "description": "Please, check your email"
          }
        }
      },
      "put": {
        "summary": "update password after forgot",
        "description": "the password must be different from the old passwords",
        "operationId": "updateForgotPassword",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Authorization",
            "type": "string",
            "required": true,
            "description": "action token"
          },
          {
            "in": "formData",
            "name": "password",
            "type": "string",
            "required": true,
            "pattern": "/^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&_])[A-Za-z\\d@$!%*?&_]{8,}$/"
          }
        ],
        "responses": {
          "200": {
            "description": "password updated"
          },
          "401": {
            "description": "Token invalid / No token"
          },
          "404": {
            "description": "Password invalid"
          },
          "409": {
            "description": "This is old password"
          }
        }
      }
    },
...
  },
...
  "securityDefinitions": {
    "accessToken": {
      "type": "apiKey",
      "name": "accessToken",
      "in": "header"
    }
  }
}